+++ docs/assets/js/search.js
@@ -1,1 +1,1 @@
-window.searchData = {"kinds":{"1":"Module","64":"Function","256":"Interface","1024":"Property","65536":"Type literal","4194304":"Type alias"},"rows":[{"id":0,"kind":1,"name":"createReactBindings","url":"modules/createreactbindings.html","classes":"tsd-kind-module"},{"id":1,"kind":64,"name":"createReactBindings","url":"modules/createreactbindings.html#createreactbindings-1","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"createReactBindings"},{"id":2,"kind":1,"name":"index","url":"modules/index.html","classes":"tsd-kind-module"},{"id":3,"kind":1,"name":"selectors","url":"modules/selectors.html","classes":"tsd-kind-module"},{"id":4,"kind":256,"name":"Selector","url":"interfaces/selectors.selector.html","classes":"tsd-kind-interface tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":5,"kind":4194304,"name":"SelectorResult","url":"modules/selectors.html#selectorresult","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":6,"kind":4194304,"name":"SelectorResults","url":"modules/selectors.html#selectorresults","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":7,"kind":4194304,"name":"LiftToSelector","url":"modules/selectors.html#lifttoselector","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":8,"kind":64,"name":"memoizeLastResult","url":"modules/selectors.html#memoizelastresult","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":9,"kind":64,"name":"createSelector","url":"modules/selectors.html#createselector","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":10,"kind":64,"name":"createCompoundSelector","url":"modules/selectors.html#createcompoundselector","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":11,"kind":1,"name":"store","url":"modules/store.html","classes":"tsd-kind-module"},{"id":12,"kind":64,"name":"createStore","url":"modules/store.html#createstore","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"store"},{"id":13,"kind":64,"name":"makeDefaultLogger","url":"modules/store.html#makedefaultlogger","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"store"},{"id":14,"kind":64,"name":"reducerFromHandlers","url":"modules/store.html#reducerfromhandlers","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"store"},{"id":15,"kind":64,"name":"createNoopLogger","url":"modules/store.html#createnooplogger","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"store"},{"id":16,"kind":1,"name":"types","url":"modules/types.html","classes":"tsd-kind-module"},{"id":17,"kind":4194304,"name":"Reducer","url":"modules/types.html#reducer","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":18,"kind":65536,"name":"__type","url":"modules/types.html#reducer.__type-9","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Reducer"},{"id":19,"kind":4194304,"name":"ActionHandlers","url":"modules/types.html#actionhandlers","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":20,"kind":4194304,"name":"GetState","url":"modules/types.html#getstate","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":21,"kind":65536,"name":"__type","url":"modules/types.html#getstate.__type-3","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.GetState"},{"id":22,"kind":4194304,"name":"Subscriber","url":"modules/types.html#subscriber","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":23,"kind":65536,"name":"__type","url":"modules/types.html#subscriber.__type-14","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Subscriber"},{"id":24,"kind":4194304,"name":"Dispatcher","url":"modules/types.html#dispatcher","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":25,"kind":65536,"name":"__type","url":"modules/types.html#dispatcher.__type-2","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Dispatcher"},{"id":26,"kind":4194304,"name":"Action","url":"modules/types.html#action","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"types"},{"id":27,"kind":65536,"name":"__type","url":"modules/types.html#action.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Action"},{"id":28,"kind":1024,"name":"type","url":"modules/types.html#action.__type.type","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Action.__type"},{"id":29,"kind":4194304,"name":"ActionThunk","url":"modules/types.html#actionthunk","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":30,"kind":65536,"name":"__type","url":"modules/types.html#actionthunk.__type-1","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.ActionThunk"},{"id":31,"kind":4194304,"name":"Logger","url":"modules/types.html#logger","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":32,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Logger"},{"id":33,"kind":1024,"name":"logStart","url":"modules/types.html#logger.__type-4.logstart","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":34,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4.__type-7","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":35,"kind":1024,"name":"logAction","url":"modules/types.html#logger.__type-4.logaction","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":36,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4.__type-5","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":37,"kind":1024,"name":"logState","url":"modules/types.html#logger.__type-4.logstate","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":38,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4.__type-8","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":39,"kind":1024,"name":"logEnd","url":"modules/types.html#logger.__type-4.logend","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":40,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4.__type-6","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":41,"kind":4194304,"name":"Store","url":"modules/types.html#store","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":42,"kind":65536,"name":"__type","url":"modules/types.html#store.__type-10","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Store"},{"id":43,"kind":1024,"name":"getState","url":"modules/types.html#store.__type-10.getstate-1","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":44,"kind":1024,"name":"dispatch","url":"modules/types.html#store.__type-10.dispatch","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":45,"kind":1024,"name":"subscribe","url":"modules/types.html#store.__type-10.subscribe","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":46,"kind":65536,"name":"__type","url":"modules/types.html#store.__type-10.__type-13","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":47,"kind":1024,"name":"addReducer","url":"modules/types.html#store.__type-10.addreducer","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":48,"kind":65536,"name":"__type","url":"modules/types.html#store.__type-10.__type-11","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":49,"kind":1024,"name":"addSubReducer","url":"modules/types.html#store.__type-10.addsubreducer","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":50,"kind":65536,"name":"__type","url":"modules/types.html#store.__type-10.__type-12","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal tsd-has-type-parameter","parent":"types.Store.__type"},{"id":51,"kind":64,"name":"createStore","url":"modules/index.html#createstore","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":52,"kind":64,"name":"makeDefaultLogger","url":"modules/index.html#makedefaultlogger","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":53,"kind":64,"name":"reducerFromHandlers","url":"modules/index.html#reducerfromhandlers","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":54,"kind":64,"name":"createNoopLogger","url":"modules/index.html#createnooplogger","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":55,"kind":64,"name":"createReactBindings","url":"modules/index.html#createreactbindings","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":56,"kind":256,"name":"Selector","url":"interfaces/index.selector.html","classes":"tsd-kind-interface tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":57,"kind":4194304,"name":"SelectorResult","url":"modules/index.html#selectorresult","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":58,"kind":4194304,"name":"SelectorResults","url":"modules/index.html#selectorresults","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":59,"kind":4194304,"name":"LiftToSelector","url":"modules/index.html#lifttoselector","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":60,"kind":64,"name":"memoizeLastResult","url":"modules/index.html#memoizelastresult","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":61,"kind":64,"name":"createSelector","url":"modules/index.html#createselector","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":62,"kind":64,"name":"createCompoundSelector","url":"modules/index.html#createcompoundselector","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"}],"index":{"version":"2.3.9","fields":["name","parent"],"fieldVectors":[["name/0",[0,26.548]],["parent/0",[]],["name/1",[0,26.548]],["parent/1",[0,2.564]],["name/2",[1,15.562]],["parent/2",[]],["name/3",[2,20.188]],["parent/3",[]],["name/4",[3,32.426]],["parent/4",[2,1.95]],["name/5",[4,32.426]],["parent/5",[2,1.95]],["name/6",[5,32.426]],["parent/6",[2,1.95]],["name/7",[6,32.426]],["parent/7",[2,1.95]],["name/8",[7,32.426]],["parent/8",[2,1.95]],["name/9",[8,32.426]],["parent/9",[2,1.95]],["name/10",[9,32.426]],["parent/10",[2,1.95]],["name/11",[10,22.871]],["parent/11",[]],["name/12",[11,32.426]],["parent/12",[10,2.209]],["name/13",[12,32.426]],["parent/13",[10,2.209]],["name/14",[13,32.426]],["parent/14",[10,2.209]],["name/15",[14,32.426]],["parent/15",[10,2.209]],["name/16",[15,18.075]],["parent/16",[]],["name/17",[16,37.534]],["parent/17",[15,1.746]],["name/18",[17,14.18]],["parent/18",[18,3.626]],["name/19",[19,37.534]],["parent/19",[15,1.746]],["name/20",[20,32.426]],["parent/20",[15,1.746]],["name/21",[17,14.18]],["parent/21",[21,3.626]],["name/22",[22,37.534]],["parent/22",[15,1.746]],["name/23",[17,14.18]],["parent/23",[23,3.626]],["name/24",[24,37.534]],["parent/24",[15,1.746]],["name/25",[17,14.18]],["parent/25",[25,3.626]],["name/26",[26,37.534]],["parent/26",[15,1.746]],["name/27",[17,14.18]],["parent/27",[27,3.626]],["name/28",[28,37.534]],["parent/28",[29,3.626]],["name/29",[30,37.534]],["parent/29",[15,1.746]],["name/30",[17,14.18]],["parent/30",[31,3.626]],["name/31",[32,37.534]],["parent/31",[15,1.746]],["name/32",[17,14.18]],["parent/32",[33,3.626]],["name/33",[34,37.534]],["parent/33",[35,1.95]],["name/34",[17,14.18]],["parent/34",[35,1.95]],["name/35",[36,37.534]],["parent/35",[35,1.95]],["name/36",[17,14.18]],["parent/36",[35,1.95]],["name/37",[37,37.534]],["parent/37",[35,1.95]],["name/38",[17,14.18]],["parent/38",[35,1.95]],["name/39",[38,37.534]],["parent/39",[35,1.95]],["name/40",[17,14.18]],["parent/40",[35,1.95]],["name/41",[10,22.871]],["parent/41",[15,1.746]],["name/42",[17,14.18]],["parent/42",[39,3.626]],["name/43",[20,32.426]],["parent/43",[40,1.95]],["name/44",[41,37.534]],["parent/44",[40,1.95]],["name/45",[42,37.534]],["parent/45",[40,1.95]],["name/46",[17,14.18]],["parent/46",[40,1.95]],["name/47",[43,37.534]],["parent/47",[40,1.95]],["name/48",[17,14.18]],["parent/48",[40,1.95]],["name/49",[44,37.534]],["parent/49",[40,1.95]],["name/50",[17,14.18]],["parent/50",[40,1.95]],["name/51",[11,32.426]],["parent/51",[1,1.503]],["name/52",[12,32.426]],["parent/52",[1,1.503]],["name/53",[13,32.426]],["parent/53",[1,1.503]],["name/54",[14,32.426]],["parent/54",[1,1.503]],["name/55",[0,26.548]],["parent/55",[1,1.503]],["name/56",[3,32.426]],["parent/56",[1,1.503]],["name/57",[4,32.426]],["parent/57",[1,1.503]],["name/58",[5,32.426]],["parent/58",[1,1.503]],["name/59",[6,32.426]],["parent/59",[1,1.503]],["name/60",[7,32.426]],["parent/60",[1,1.503]],["name/61",[8,32.426]],["parent/61",[1,1.503]],["name/62",[9,32.426]],["parent/62",[1,1.503]]],"invertedIndex":[["__type",{"_index":17,"name":{"18":{},"21":{},"23":{},"25":{},"27":{},"30":{},"32":{},"34":{},"36":{},"38":{},"40":{},"42":{},"46":{},"48":{},"50":{}},"parent":{}}],["action",{"_index":26,"name":{"26":{}},"parent":{}}],["actionhandlers",{"_index":19,"name":{"19":{}},"parent":{}}],["actionthunk",{"_index":30,"name":{"29":{}},"parent":{}}],["addreducer",{"_index":43,"name":{"47":{}},"parent":{}}],["addsubreducer",{"_index":44,"name":{"49":{}},"parent":{}}],["createcompoundselector",{"_index":9,"name":{"10":{},"62":{}},"parent":{}}],["createnooplogger",{"_index":14,"name":{"15":{},"54":{}},"parent":{}}],["createreactbindings",{"_index":0,"name":{"0":{},"1":{},"55":{}},"parent":{"1":{}}}],["createselector",{"_index":8,"name":{"9":{},"61":{}},"parent":{}}],["createstore",{"_index":11,"name":{"12":{},"51":{}},"parent":{}}],["dispatch",{"_index":41,"name":{"44":{}},"parent":{}}],["dispatcher",{"_index":24,"name":{"24":{}},"parent":{}}],["getstate",{"_index":20,"name":{"20":{},"43":{}},"parent":{}}],["index",{"_index":1,"name":{"2":{}},"parent":{"51":{},"52":{},"53":{},"54":{},"55":{},"56":{},"57":{},"58":{},"59":{},"60":{},"61":{},"62":{}}}],["lifttoselector",{"_index":6,"name":{"7":{},"59":{}},"parent":{}}],["logaction",{"_index":36,"name":{"35":{}},"parent":{}}],["logend",{"_index":38,"name":{"39":{}},"parent":{}}],["logger",{"_index":32,"name":{"31":{}},"parent":{}}],["logstart",{"_index":34,"name":{"33":{}},"parent":{}}],["logstate",{"_index":37,"name":{"37":{}},"parent":{}}],["makedefaultlogger",{"_index":12,"name":{"13":{},"52":{}},"parent":{}}],["memoizelastresult",{"_index":7,"name":{"8":{},"60":{}},"parent":{}}],["reducer",{"_index":16,"name":{"17":{}},"parent":{}}],["reducerfromhandlers",{"_index":13,"name":{"14":{},"53":{}},"parent":{}}],["selector",{"_index":3,"name":{"4":{},"56":{}},"parent":{}}],["selectorresult",{"_index":4,"name":{"5":{},"57":{}},"parent":{}}],["selectorresults",{"_index":5,"name":{"6":{},"58":{}},"parent":{}}],["selectors",{"_index":2,"name":{"3":{}},"parent":{"4":{},"5":{},"6":{},"7":{},"8":{},"9":{},"10":{}}}],["store",{"_index":10,"name":{"11":{},"41":{}},"parent":{"12":{},"13":{},"14":{},"15":{}}}],["subscribe",{"_index":42,"name":{"45":{}},"parent":{}}],["subscriber",{"_index":22,"name":{"22":{}},"parent":{}}],["type",{"_index":28,"name":{"28":{}},"parent":{}}],["types",{"_index":15,"name":{"16":{}},"parent":{"17":{},"19":{},"20":{},"22":{},"24":{},"26":{},"29":{},"31":{},"41":{}}}],["types.action",{"_index":27,"name":{},"parent":{"27":{}}}],["types.action.__type",{"_index":29,"name":{},"parent":{"28":{}}}],["types.actionthunk",{"_index":31,"name":{},"parent":{"30":{}}}],["types.dispatcher",{"_index":25,"name":{},"parent":{"25":{}}}],["types.getstate",{"_index":21,"name":{},"parent":{"21":{}}}],["types.logger",{"_index":33,"name":{},"parent":{"32":{}}}],["types.logger.__type",{"_index":35,"name":{},"parent":{"33":{},"34":{},"35":{},"36":{},"37":{},"38":{},"39":{},"40":{}}}],["types.reducer",{"_index":18,"name":{},"parent":{"18":{}}}],["types.store",{"_index":39,"name":{},"parent":{"42":{}}}],["types.store.__type",{"_index":40,"name":{},"parent":{"43":{},"44":{},"45":{},"46":{},"47":{},"48":{},"49":{},"50":{}}}],["types.subscriber",{"_index":23,"name":{},"parent":{"23":{}}}]],"pipeline":[]}}
+window.searchData = {"kinds":{"1":"Module","64":"Function","256":"Interface","1024":"Property","65536":"Type literal","4194304":"Type alias"},"rows":[{"id":0,"kind":1,"name":"createReactBindings","url":"modules/createreactbindings.html","classes":"tsd-kind-module"},{"id":1,"kind":64,"name":"createReactBindings","url":"modules/createreactbindings.html#createreactbindings-1","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"createReactBindings"},{"id":2,"kind":1,"name":"index","url":"modules/index.html","classes":"tsd-kind-module"},{"id":3,"kind":1,"name":"selectors","url":"modules/selectors.html","classes":"tsd-kind-module"},{"id":4,"kind":64,"name":"createSelector","url":"modules/selectors.html#createselector","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":5,"kind":64,"name":"createCompoundSelector","url":"modules/selectors.html#createcompoundselector","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"selectors"},{"id":6,"kind":1,"name":"store","url":"modules/store.html","classes":"tsd-kind-module"},{"id":7,"kind":64,"name":"createStore","url":"modules/store.html#createstore","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"store"},{"id":8,"kind":64,"name":"makeDefaultLogger","url":"modules/store.html#makedefaultlogger","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"store"},{"id":9,"kind":64,"name":"reducerFromHandlers","url":"modules/store.html#reducerfromhandlers","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"store"},{"id":10,"kind":64,"name":"createNoopLogger","url":"modules/store.html#createnooplogger","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"store"},{"id":11,"kind":1,"name":"types","url":"modules/types.html","classes":"tsd-kind-module"},{"id":12,"kind":4194304,"name":"Reducer","url":"modules/types.html#reducer","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":13,"kind":65536,"name":"__type","url":"modules/types.html#reducer.__type-10","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Reducer"},{"id":14,"kind":4194304,"name":"ActionHandlers","url":"modules/types.html#actionhandlers","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":15,"kind":4194304,"name":"GetState","url":"modules/types.html#getstate","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":16,"kind":65536,"name":"__type","url":"modules/types.html#getstate.__type-3","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.GetState"},{"id":17,"kind":4194304,"name":"Subscriber","url":"modules/types.html#subscriber","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":18,"kind":65536,"name":"__type","url":"modules/types.html#subscriber.__type-15","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Subscriber"},{"id":19,"kind":4194304,"name":"Dispatcher","url":"modules/types.html#dispatcher","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":20,"kind":65536,"name":"__type","url":"modules/types.html#dispatcher.__type-2","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Dispatcher"},{"id":21,"kind":4194304,"name":"Action","url":"modules/types.html#action","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"types"},{"id":22,"kind":65536,"name":"__type","url":"modules/types.html#action.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Action"},{"id":23,"kind":1024,"name":"type","url":"modules/types.html#action.__type.type","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Action.__type"},{"id":24,"kind":4194304,"name":"ActionThunk","url":"modules/types.html#actionthunk","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":25,"kind":65536,"name":"__type","url":"modules/types.html#actionthunk.__type-1","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.ActionThunk"},{"id":26,"kind":4194304,"name":"Logger","url":"modules/types.html#logger","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":27,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Logger"},{"id":28,"kind":1024,"name":"logStart","url":"modules/types.html#logger.__type-4.logstart","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":29,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4.__type-7","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":30,"kind":1024,"name":"logAction","url":"modules/types.html#logger.__type-4.logaction","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":31,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4.__type-5","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":32,"kind":1024,"name":"logState","url":"modules/types.html#logger.__type-4.logstate","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":33,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4.__type-8","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":34,"kind":1024,"name":"logEnd","url":"modules/types.html#logger.__type-4.logend","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":35,"kind":65536,"name":"__type","url":"modules/types.html#logger.__type-4.__type-6","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Logger.__type"},{"id":36,"kind":4194304,"name":"Store","url":"modules/types.html#store","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":37,"kind":65536,"name":"__type","url":"modules/types.html#store.__type-11","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias","parent":"types.Store"},{"id":38,"kind":1024,"name":"getState","url":"modules/types.html#store.__type-11.getstate-1","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":39,"kind":1024,"name":"dispatch","url":"modules/types.html#store.__type-11.dispatch","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":40,"kind":1024,"name":"subscribe","url":"modules/types.html#store.__type-11.subscribe","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":41,"kind":65536,"name":"__type","url":"modules/types.html#store.__type-11.__type-14","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":42,"kind":1024,"name":"addReducer","url":"modules/types.html#store.__type-11.addreducer","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":43,"kind":65536,"name":"__type","url":"modules/types.html#store.__type-11.__type-12","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":44,"kind":1024,"name":"addSubReducer","url":"modules/types.html#store.__type-11.addsubreducer","classes":"tsd-kind-property tsd-parent-kind-type-literal","parent":"types.Store.__type"},{"id":45,"kind":65536,"name":"__type","url":"modules/types.html#store.__type-11.__type-13","classes":"tsd-kind-type-literal tsd-parent-kind-type-literal tsd-has-type-parameter","parent":"types.Store.__type"},{"id":46,"kind":256,"name":"Selector","url":"interfaces/types.selector.html","classes":"tsd-kind-interface tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":47,"kind":4194304,"name":"SelectorResult","url":"modules/types.html#selectorresult","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":48,"kind":4194304,"name":"SelectorResults","url":"modules/types.html#selectorresults","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":49,"kind":4194304,"name":"LiftToSelector","url":"modules/types.html#lifttoselector","classes":"tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter","parent":"types"},{"id":50,"kind":4194304,"name":"Memoize","url":"modules/types.html#memoize","classes":"tsd-kind-type-alias tsd-parent-kind-module","parent":"types"},{"id":51,"kind":65536,"name":"__type","url":"modules/types.html#memoize.__type-9","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-has-type-parameter","parent":"types.Memoize"},{"id":52,"kind":64,"name":"createStore","url":"modules/index.html#createstore","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":53,"kind":64,"name":"makeDefaultLogger","url":"modules/index.html#makedefaultlogger","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":54,"kind":64,"name":"reducerFromHandlers","url":"modules/index.html#reducerfromhandlers","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":55,"kind":64,"name":"createNoopLogger","url":"modules/index.html#createnooplogger","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":56,"kind":64,"name":"createReactBindings","url":"modules/index.html#createreactbindings","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":57,"kind":64,"name":"createSelector","url":"modules/index.html#createselector","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"},{"id":58,"kind":64,"name":"createCompoundSelector","url":"modules/index.html#createcompoundselector","classes":"tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter","parent":"index"}],"index":{"version":"2.3.9","fields":["name","parent"],"fieldVectors":[["name/0",[0,25.903]],["parent/0",[]],["name/1",[0,25.903]],["parent/1",[0,2.496]],["name/2",[1,19.543]],["parent/2",[]],["name/3",[2,28.416]],["parent/3",[]],["name/4",[3,31.781]],["parent/4",[2,2.738]],["name/5",[4,31.781]],["parent/5",[2,2.738]],["name/6",[5,22.225]],["parent/6",[]],["name/7",[6,31.781]],["parent/7",[5,2.141]],["name/8",[7,31.781]],["parent/8",[5,2.141]],["name/9",[8,31.781]],["parent/9",[5,2.141]],["name/10",[9,31.781]],["parent/10",[5,2.141]],["name/11",[10,13.535]],["parent/11",[]],["name/12",[11,36.889]],["parent/12",[10,1.304]],["name/13",[12,12.91]],["parent/13",[13,3.554]],["name/14",[14,36.889]],["parent/14",[10,1.304]],["name/15",[15,31.781]],["parent/15",[10,1.304]],["name/16",[12,12.91]],["parent/16",[16,3.554]],["name/17",[17,36.889]],["parent/17",[10,1.304]],["name/18",[12,12.91]],["parent/18",[18,3.554]],["name/19",[19,36.889]],["parent/19",[10,1.304]],["name/20",[12,12.91]],["parent/20",[20,3.554]],["name/21",[21,36.889]],["parent/21",[10,1.304]],["name/22",[12,12.91]],["parent/22",[22,3.554]],["name/23",[23,36.889]],["parent/23",[24,3.554]],["name/24",[25,36.889]],["parent/24",[10,1.304]],["name/25",[12,12.91]],["parent/25",[26,3.554]],["name/26",[27,36.889]],["parent/26",[10,1.304]],["name/27",[12,12.91]],["parent/27",[28,3.554]],["name/28",[29,36.889]],["parent/28",[30,1.883]],["name/29",[12,12.91]],["parent/29",[30,1.883]],["name/30",[31,36.889]],["parent/30",[30,1.883]],["name/31",[12,12.91]],["parent/31",[30,1.883]],["name/32",[32,36.889]],["parent/32",[30,1.883]],["name/33",[12,12.91]],["parent/33",[30,1.883]],["name/34",[33,36.889]],["parent/34",[30,1.883]],["name/35",[12,12.91]],["parent/35",[30,1.883]],["name/36",[5,22.225]],["parent/36",[10,1.304]],["name/37",[12,12.91]],["parent/37",[34,3.554]],["name/38",[15,31.781]],["parent/38",[35,1.883]],["name/39",[36,36.889]],["parent/39",[35,1.883]],["name/40",[37,36.889]],["parent/40",[35,1.883]],["name/41",[12,12.91]],["parent/41",[35,1.883]],["name/42",[38,36.889]],["parent/42",[35,1.883]],["name/43",[12,12.91]],["parent/43",[35,1.883]],["name/44",[39,36.889]],["parent/44",[35,1.883]],["name/45",[12,12.91]],["parent/45",[35,1.883]],["name/46",[40,36.889]],["parent/46",[10,1.304]],["name/47",[41,36.889]],["parent/47",[10,1.304]],["name/48",[42,36.889]],["parent/48",[10,1.304]],["name/49",[43,36.889]],["parent/49",[10,1.304]],["name/50",[44,36.889]],["parent/50",[10,1.304]],["name/51",[12,12.91]],["parent/51",[45,3.554]],["name/52",[6,31.781]],["parent/52",[1,1.883]],["name/53",[7,31.781]],["parent/53",[1,1.883]],["name/54",[8,31.781]],["parent/54",[1,1.883]],["name/55",[9,31.781]],["parent/55",[1,1.883]],["name/56",[0,25.903]],["parent/56",[1,1.883]],["name/57",[3,31.781]],["parent/57",[1,1.883]],["name/58",[4,31.781]],["parent/58",[1,1.883]]],"invertedIndex":[["__type",{"_index":12,"name":{"13":{},"16":{},"18":{},"20":{},"22":{},"25":{},"27":{},"29":{},"31":{},"33":{},"35":{},"37":{},"41":{},"43":{},"45":{},"51":{}},"parent":{}}],["action",{"_index":21,"name":{"21":{}},"parent":{}}],["actionhandlers",{"_index":14,"name":{"14":{}},"parent":{}}],["actionthunk",{"_index":25,"name":{"24":{}},"parent":{}}],["addreducer",{"_index":38,"name":{"42":{}},"parent":{}}],["addsubreducer",{"_index":39,"name":{"44":{}},"parent":{}}],["createcompoundselector",{"_index":4,"name":{"5":{},"58":{}},"parent":{}}],["createnooplogger",{"_index":9,"name":{"10":{},"55":{}},"parent":{}}],["createreactbindings",{"_index":0,"name":{"0":{},"1":{},"56":{}},"parent":{"1":{}}}],["createselector",{"_index":3,"name":{"4":{},"57":{}},"parent":{}}],["createstore",{"_index":6,"name":{"7":{},"52":{}},"parent":{}}],["dispatch",{"_index":36,"name":{"39":{}},"parent":{}}],["dispatcher",{"_index":19,"name":{"19":{}},"parent":{}}],["getstate",{"_index":15,"name":{"15":{},"38":{}},"parent":{}}],["index",{"_index":1,"name":{"2":{}},"parent":{"52":{},"53":{},"54":{},"55":{},"56":{},"57":{},"58":{}}}],["lifttoselector",{"_index":43,"name":{"49":{}},"parent":{}}],["logaction",{"_index":31,"name":{"30":{}},"parent":{}}],["logend",{"_index":33,"name":{"34":{}},"parent":{}}],["logger",{"_index":27,"name":{"26":{}},"parent":{}}],["logstart",{"_index":29,"name":{"28":{}},"parent":{}}],["logstate",{"_index":32,"name":{"32":{}},"parent":{}}],["makedefaultlogger",{"_index":7,"name":{"8":{},"53":{}},"parent":{}}],["memoize",{"_index":44,"name":{"50":{}},"parent":{}}],["reducer",{"_index":11,"name":{"12":{}},"parent":{}}],["reducerfromhandlers",{"_index":8,"name":{"9":{},"54":{}},"parent":{}}],["selector",{"_index":40,"name":{"46":{}},"parent":{}}],["selectorresult",{"_index":41,"name":{"47":{}},"parent":{}}],["selectorresults",{"_index":42,"name":{"48":{}},"parent":{}}],["selectors",{"_index":2,"name":{"3":{}},"parent":{"4":{},"5":{}}}],["store",{"_index":5,"name":{"6":{},"36":{}},"parent":{"7":{},"8":{},"9":{},"10":{}}}],["subscribe",{"_index":37,"name":{"40":{}},"parent":{}}],["subscriber",{"_index":17,"name":{"17":{}},"parent":{}}],["type",{"_index":23,"name":{"23":{}},"parent":{}}],["types",{"_index":10,"name":{"11":{}},"parent":{"12":{},"14":{},"15":{},"17":{},"19":{},"21":{},"24":{},"26":{},"36":{},"46":{},"47":{},"48":{},"49":{},"50":{}}}],["types.action",{"_index":22,"name":{},"parent":{"22":{}}}],["types.action.__type",{"_index":24,"name":{},"parent":{"23":{}}}],["types.actionthunk",{"_index":26,"name":{},"parent":{"25":{}}}],["types.dispatcher",{"_index":20,"name":{},"parent":{"20":{}}}],["types.getstate",{"_index":16,"name":{},"parent":{"16":{}}}],["types.logger",{"_index":28,"name":{},"parent":{"27":{}}}],["types.logger.__type",{"_index":30,"name":{},"parent":{"28":{},"29":{},"30":{},"31":{},"32":{},"33":{},"34":{},"35":{}}}],["types.memoize",{"_index":45,"name":{},"parent":{"51":{}}}],["types.reducer",{"_index":13,"name":{},"parent":{"13":{}}}],["types.store",{"_index":34,"name":{},"parent":{"37":{}}}],["types.store.__type",{"_index":35,"name":{},"parent":{"38":{},"39":{},"40":{},"41":{},"42":{},"43":{},"44":{},"45":{}}}],["types.subscriber",{"_index":18,"name":{},"parent":{"18":{}}}]],"pipeline":[]}}
