+++ docs/interfaces/types.selector.html
@@ -55,10 +55,10 @@
 					<a href="../modules.html">topstate</a>
 				</li>
 				<li>
-					<a href="../modules/index.html">index</a>
+					<a href="../modules/types.html">types</a>
 				</li>
 				<li>
-					<a href="index.selector.html">Selector</a>
+					<a href="types.selector.html">Selector</a>
 				</li>
 			</ul>
 			<h1>Interface Selector&lt;S, A&gt;</h1>
@@ -96,7 +96,7 @@
 					<li class="tsd-description">
 						<aside class="tsd-sources">
 							<ul>
-								<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/157f55e/src/selectors.ts#L1">selectors.ts:1</a></li>
+								<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/1e85bbc/src/types.ts#L32">types.ts:32</a></li>
 							</ul>
 						</aside>
 						<h4 class="tsd-parameters-title">Parameters</h4>
@@ -119,7 +119,7 @@
 					<li class=" tsd-kind-module">
 						<a href="../modules/createreactbindings.html">create<wbr>React<wbr>Bindings</a>
 					</li>
-					<li class="current tsd-kind-module">
+					<li class=" tsd-kind-module">
 						<a href="../modules/index.html">index</a>
 					</li>
 					<li class=" tsd-kind-module">
@@ -128,7 +128,7 @@
 					<li class=" tsd-kind-module">
 						<a href="../modules/store.html">store</a>
 					</li>
-					<li class=" tsd-kind-module">
+					<li class="current tsd-kind-module">
 						<a href="../modules/types.html">types</a>
 					</li>
 				</ul>
@@ -138,42 +138,48 @@
 				</ul>
 				<ul class="current">
 					<li class="current tsd-kind-interface tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="index.selector.html" class="tsd-kind-icon">Selector</a>
+						<a href="types.selector.html" class="tsd-kind-icon">Selector</a>
 					</li>
 				</ul>
 				<ul class="after-current">
-					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#lifttoselector" class="tsd-kind-icon">Lift<wbr>ToSelector</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module">
+						<a href="../modules/types.html#action" class="tsd-kind-icon">Action</a>
 					</li>
 					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#selectorresult" class="tsd-kind-icon">Selector<wbr>Result</a>
+						<a href="../modules/types.html#actionhandlers" class="tsd-kind-icon">Action<wbr>Handlers</a>
 					</li>
 					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#selectorresults" class="tsd-kind-icon">Selector<wbr>Results</a>
+						<a href="../modules/types.html#actionthunk" class="tsd-kind-icon">Action<wbr>Thunk</a>
 					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#createcompoundselector" class="tsd-kind-icon">create<wbr>Compound<wbr>Selector</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#dispatcher" class="tsd-kind-icon">Dispatcher</a>
 					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#createnooplogger" class="tsd-kind-icon">create<wbr>Noop<wbr>Logger</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#getstate" class="tsd-kind-icon">Get<wbr>State</a>
 					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#createreactbindings" class="tsd-kind-icon">create<wbr>React<wbr>Bindings</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#lifttoselector" class="tsd-kind-icon">Lift<wbr>ToSelector</a>
 					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#createselector" class="tsd-kind-icon">create<wbr>Selector</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#logger" class="tsd-kind-icon">Logger</a>
 					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#createstore" class="tsd-kind-icon">create<wbr>Store</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module">
+						<a href="../modules/types.html#memoize" class="tsd-kind-icon">Memoize</a>
 					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#makedefaultlogger" class="tsd-kind-icon">make<wbr>Default<wbr>Logger</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#reducer" class="tsd-kind-icon">Reducer</a>
 					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#memoizelastresult" class="tsd-kind-icon">memoize<wbr>Last<wbr>Result</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#selectorresult" class="tsd-kind-icon">Selector<wbr>Result</a>
 					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-module tsd-has-type-parameter">
-						<a href="../modules/index.html#reducerfromhandlers" class="tsd-kind-icon">reducer<wbr>From<wbr>Handlers</a>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#selectorresults" class="tsd-kind-icon">Selector<wbr>Results</a>
+					</li>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#store" class="tsd-kind-icon">Store</a>
+					</li>
+					<li class=" tsd-kind-type-alias tsd-parent-kind-module tsd-has-type-parameter">
+						<a href="../modules/types.html#subscriber" class="tsd-kind-icon">Subscriber</a>
 					</li>
 				</ul>
 			</nav>
