+++ docs/modules/store.html
@@ -93,7 +93,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/157f55e/src/store.ts#L142">store.ts:142</a></li>
+									<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/1e85bbc/src/store.ts#L142">store.ts:142</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -124,7 +124,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/157f55e/src/store.ts#L25">store.ts:25</a></li>
+									<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/1e85bbc/src/store.ts#L25">store.ts:25</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -170,7 +170,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/157f55e/src/store.ts#L13">store.ts:13</a></li>
+									<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/1e85bbc/src/store.ts#L13">store.ts:13</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-type-parameters-title">Type parameters</h4>
@@ -196,7 +196,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/157f55e/src/store.ts#L128">store.ts:128</a></li>
+									<li>Defined in <a href="https://github.com/Willmo36/topstate/blob/1e85bbc/src/store.ts#L128">store.ts:128</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
